chore: mixed up tests- fix
